+++ src/commonMain/kotlin/org/csenseoss/kotlin/annotations/numbers/limit/DoubleLimit.kt
@@ -1,0 +1,21 @@
+@file:Suppress("unused")
+
+package org.csenseoss.kotlin.annotations.numbers.limit
+
+@Target(
+    AnnotationTarget.FUNCTION,
+    AnnotationTarget.PROPERTY,
+    AnnotationTarget.PROPERTY_GETTER,
+    AnnotationTarget.PROPERTY_SETTER,
+    AnnotationTarget.VALUE_PARAMETER,
+    AnnotationTarget.FIELD,
+    AnnotationTarget.LOCAL_VARIABLE,
+    AnnotationTarget.ANNOTATION_CLASS,
+    AnnotationTarget.TYPE//until you can annotate functional parameters
+)
+@Retention(AnnotationRetention.BINARY)
+@MustBeDocumented
+public annotation class DoubleLimit(
+    val from: Double = Double.MIN_VALUE,
+    val to: Double = Double.MAX_VALUE
+)