+++ src/commonMain/kotlin/org/csenseoss/kotlin/annotations/threading/in/InIO.kt
@@ -1,0 +1,16 @@
+package org.csenseoss.kotlin.annotations.threading.`in`
+
+/**
+ * Denotes that the given MUST be executed in the IO thread
+ */
+@Target(
+    AnnotationTarget.FUNCTION,
+    AnnotationTarget.PROPERTY,
+    AnnotationTarget.PROPERTY_SETTER,
+    AnnotationTarget.PROPERTY_GETTER,
+    AnnotationTarget.VALUE_PARAMETER,
+    AnnotationTarget.TYPE //for functional types
+)
+@Retention(AnnotationRetention.BINARY)
+@MustBeDocumented
+public annotation class InIO