+++ src/commonMain/kotlin/org/csenseoss/kotlin/annotations/threading/in/InIOContext.kt
@@ -1,0 +1,14 @@
+package org.csenseoss.kotlin.annotations.threading.`in`
+
+/**
+ * Denotes that the given MUST be executed in the IO thread
+ * This applies to the whole scope (eg a whole class)
+ */
+@Retention(AnnotationRetention.BINARY)
+@Target(
+    AnnotationTarget.CLASS,
+    AnnotationTarget.FILE,
+    AnnotationTarget.TYPE
+)
+@MustBeDocumented
+public annotation class InIOContext