+++ src/commonMain/kotlin/org/csenseoss/kotlin/annotations/threading/in/InUi.kt
@@ -1,0 +1,19 @@
+@file:Suppress("unused")
+
+package org.csenseoss.kotlin.annotations.threading.`in`
+
+
+/**
+ * Denotes that the given MUST be executed in the UI thread
+ */
+@Target(
+    AnnotationTarget.FUNCTION,
+    AnnotationTarget.PROPERTY,
+    AnnotationTarget.PROPERTY_SETTER,
+    AnnotationTarget.PROPERTY_GETTER,
+    AnnotationTarget.VALUE_PARAMETER,
+    AnnotationTarget.TYPE //for functional types
+)
+@Retention(AnnotationRetention.BINARY) 
+@MustBeDocumented
+public annotation class InUi