+++ changelog.md
@@ -1,3 +1,16 @@
+# 1.0.0
+- kotlin 2.1.21
+- rename package (for maven central)
+- added
+  - InIOContext
+  - MustImplementEquals
+  - UseSingletonInstance
+
+- first version of "UnitSystem"
+- updated namespaces for "in" (threading), "size" (collection) limit (numbers)
+- added "ColoredCode"
+- added InIO (akin to inBackground / inUI)
+
 # 0.0.63
 
 - fixed targets for CatchesException(s) & RethrowsExceptions